Phase 2.6 (cont): handle tables across modules + amxmodx core
Continues the per-module cell↔native-pointer audit. All sites use
the PtrHandleTable<T> facility from public/AMXModulePtrHandle.h so
plugins get a 1-based opaque cell handle instead of a truncated
native pointer.

amxmodx core
- file.cpp: g_file_handles / g_dir_handles for FileObject* and
  DirectoryHandle*. fopen/freadX/fclose/open_dir/next_file/close_dir
  all routed through the table; fclose / close_dir release the slot.
- gameconfigs.cpp: g_gameconf_addr_handles for GameConfGetAddress()
  return values.
- cvars.cpp: g_cvar_handles + g_cvarhook_handles for cvar_t* and
  AutoForward*. create_cvar / register_cvar / get_cvar_pointer /
  hook_cvar_change / enable_cvar_hook / disable_cvar_hook etc.

fakemeta module
- New fm_handles.h: shared PtrHandleTable<void> facility
  (g_fm_ptr_handles) plus fm_ptr_to_cell / fm_cell_to_ptr<T> helpers.
  Single global table owned by fakemeta_amxx.cpp.
- fm_tr2.cpp: migrated from per-type local tables (g_tr_handles,
  g_kvd_handles) to the shared one. set_tr2 / get_tr2 / set_kvd /
  get_kvd / get_cd / set_cd / get_es / set_es / get_uc / set_uc all
  resolve params[1] via fm_cell_to_ptr<T>.
- forward.cpp: every (cell)pointer cast in the engine forward
  emit-sites (TraceResult*, clientdata*, entity_state*, usercmd*,
  KeyValueData*, set_t* pSet, char* infobuffer/data) replaced with
  fm_ptr_to_cell. GetInfoKeyBuffer return value round-trips
  correctly via fm_cell_to_ptr<char>.
- engfunc.cpp:948,1081 — pset cell unmarshalled via
  pdata_cell_to_pointer (HLTypeConversion's table); GetInfoKeyBuffer
  return marshalled via pdata_pointer_to_cell.

ns module
- particles.cpp: edict_t* arguments switched from cell-as-pointer to
  the proper INDEXENT/ENTINDEX index-as-cell convention. Fixes
  ns_create_partsys / ns_spawn_ps / ns_keyvalue family. No handle
  table needed — entities have stable indices in the engine.

amxmodx core / compiler
- compiler/amxxpc/amx.cpp:OP_SYSREQ_D — interpreter path's
  ((AMX_NATIVE)offs)(...) call now goes through uintptr_t intermediate
  to suppress the int-to-ptr-cast warning. The fast-path is gated off
  on 64-bit in amx_Callback (Phase 2.1) so this code never runs there;
  the cast just needs to not break the compile.
- compiler/libpc300/sc1.c: the PAWN_CELL_SIZE==32 cellmax/cellmin
  constants used LONG_MAX/LONG_MIN, which on 64-bit Linux is 2^63-1
  and overflows when narrowed to a 32-bit cell. Switched to
  INT32_MAX/INT32_MIN which are unambiguously 32-bit.

Build state after this commit: aarch64 cross-build now blocked on
modules/hamsandwich/Trampolines.h:614 — the hand-rolled x86 bytecode
trampoline emitter, which is the next major piece (Phase 2.2,
AsmJit-based rewrite).

+++ compiler/amxxpc/amx.cpp
@@ -3468,7 +3468,11 @@ int AMXAPI amx_Exec(AMX *amx, cell *retval, int index)
       amx->hea=hea;
       amx->frm=frm;
       amx->stk=stk;
-      pri=((AMX_NATIVE)offs)(amx,(cell *)(data+(int)stk));
+      /* SYSREQ.D fast-path: only valid when sizeof(void*) <= sizeof(cell);
+       * on 64-bit hosts the SYSREQ.C->SYSREQ.D patching is gated off in
+       * amx_Callback, so this path never runs there. The uintptr_t
+       * intermediate suppresses the int->ptr-cast warning. */
+      pri=((AMX_NATIVE)(uintptr_t)offs)(amx,(cell *)(data+(int)stk));
       if (amx->error!=AMX_ERR_NONE) {
         if (amx->error==AMX_ERR_SLEEP) {
           amx->pri=pri;
